requirements.txt
astroid==2.4.2
bcrypt==3.2.0
beautifulsoup4==4.9.3
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
configobj==5.0.6
cryptography==3.2.1
cx-Oracle==8.0.1
cycler==0.10.0
idna==2.10
isort==5.6.4
kafka-python==2.0.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
lxml==4.6.2
matplotlib==3.3.3
mccabe==0.6.1
numpy==1.19.4
pandas==1.1.4
paramiko==2.7.2
Pillow==8.0.1
pycparser==2.20
pylint==2.6.0
PyNaCl==1.4.0
pyparsing==2.4.7
pypiwin32==223
PyQt5==5.15.2
PyQt5-sip==12.8.1
python-dateutil==2.8.1
pytz==2020.4
pywin32==300
requests==2.25.0
schedule==0.6.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
xlrd==2.0.1
xmltodict==0.12.0
